Move charge control subcommand to new file
