+++ ectool/src/main.rs
@@ -1,6 +1,7 @@
 use std::fs::File;
 use std::str::FromStr;
 
+use charge_control_subcommand::{charge_control_subcommand, ChargeControlSubcommand};
 use clap::{Parser, Subcommand, ValueEnum};
 use color_eyre::eyre::Result;
 use crosec::commands::fp_info::fp_info;
@@ -14,15 +15,11 @@ use num_traits::cast::FromPrimitive;
 
 use crosec::battery::battery;
 use crosec::commands::board_version::ec_cmd_board_version;
-use crosec::commands::charge_control::{
-    get_charge_control, set_charge_control, supports_get_and_sustainer, Sustainer,
-};
 use crosec::commands::get_cmd_versions::ec_cmd_get_cmd_versions;
 use crosec::commands::get_features::{ec_cmd_get_features, EC_FEATURE_PWM_FAN};
 use crosec::commands::set_fan_target_rpm::ec_cmd_set_fan_target_rpm;
 use crosec::commands::{
-    charge_control, get_chip_info::ec_cmd_get_chip_info, hello::ec_cmd_hello,
-    version::ec_cmd_version, CrosEcCmd,
+    get_chip_info::ec_cmd_get_chip_info, hello::ec_cmd_hello, version::ec_cmd_version, CrosEcCmd,
 };
 use crosec::console::console;
 use crosec::get_number_of_fans::{get_number_of_fans, Error};
@@ -32,6 +29,7 @@ use crosec::{
     EC_FAN_SPEED_STALLED, EC_MEM_MAP_FAN,
 };
 
+mod charge_control_subcommand;
 mod fp_download_subcommand;
 
 #[derive(Parser)]
@@ -94,7 +92,7 @@ enum Commands {
     Battery,
     ChargeControl {
         #[command(subcommand)]
-        command: Option<ChargeControlSubcommands>,
+        command: Option<ChargeControlSubcommand>,
     },
     FpInfo,
     FpStats,
@@ -114,21 +112,6 @@ enum Commands {
         #[command(subcommand)]
         command: FpDownloadSubcommand,
     },
-}
-
-#[derive(Subcommand)]
-enum ChargeControlSubcommands {
-    /// Charge the battery with external power and power the device with external power
-    Normal {
-        /// Minimum battery % to keep the battery at
-        min_percent: Option<u8>,
-        /// Maximum battery & to keep the battery at. If this isn't specified, this will be set to the same as the min %.
-        max_percent: Option<u8>,
-    },
-    /// Power the device with external power, but do not charge the battery
-    Idle,
-    /// Power the device with the battery and do not charge the battery
-    Discharge,
 }
 
 fn main() -> Result<()> {
@@ -235,52 +218,7 @@ fn main() -> Result<()> {
             let battery_info = battery(&mut file)?;
             println!("{battery_info:#?}");
         }
-        Commands::ChargeControl { command } => {
-            let mut file = File::open(CROS_EC_PATH)?;
-            match command {
-                None => {
-                    if supports_get_and_sustainer(&mut file)? {
-                        let charge_control = get_charge_control(&mut file)?;
-                        println!("{charge_control:#?}");
-                    } else {
-                        println!("This EC doesn't support getting charge control");
-                    }
-                }
-                Some(command) => match command {
-                    ChargeControlSubcommands::Normal {
-                        min_percent,
-                        max_percent,
-                    } => match min_percent {
-                        Some(min_percent) => {
-                            let max_percent = max_percent.unwrap_or(min_percent);
-                            set_charge_control(
-                                &mut file,
-                                charge_control::ChargeControl::Normal(Some(Sustainer {
-                                    min_percent: min_percent as i8,
-                                    max_percent: max_percent as i8,
-                                })),
-                            )?;
-                            println!("Set charge control to normal with sustainer from {min_percent}% to {max_percent}%");
-                        }
-                        None => {
-                            set_charge_control(
-                                &mut file,
-                                charge_control::ChargeControl::Normal(None),
-                            )?;
-                            println!("Set charge control to normal");
-                        }
-                    },
-                    ChargeControlSubcommands::Idle => {
-                        println!("Set charge control to idle");
-                        set_charge_control(&mut file, charge_control::ChargeControl::Idle)?;
-                    }
-                    ChargeControlSubcommands::Discharge => {
-                        println!("Set charge control to discharge");
-                        set_charge_control(&mut file, charge_control::ChargeControl::Discharge)?;
-                    }
-                },
-            }
-        }
+        Commands::ChargeControl { command } => charge_control_subcommand(command)?,
         Commands::FpInfo => {
             let mut file = File::open(CROS_FP_PATH)?;
             let info = fp_info(&mut file)?;
